pyproject.toml
[project]
name = "noaises-local"
version = "0.1.0"
description = "Noaises — Local-first, voice-first AI companion powered by Claude"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "claude-agent-sdk>=0.1.21",
    "anthropic>=0.52.0",
    "tomli-w>=1.0.0",
    "faster-whisper>=1.0.0",
    "sounddevice>=0.4.0",
    "numpy>=1.24.0",
    "azure-cognitiveservices-speech>=1.35.0",
    "pywebview>=5.0.0",
]

[project.scripts]
noaises = "noaises.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/noaises"]
